fix for PADDLE_ENFORCE_NOT_NULL, tests test=develop
